Rename : 리액트 프로젝트로 변경 작업(디렉토리만 구조화 함)
