--- index.html
@@ -1,171 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- Open Graph 메타 태그 -->
-    <meta property="og:image" content="img/Img_home_03.png" />
-    <meta property="og:title" content="판다 마켓" />
-    <meta property="og:description" content="일상의 모든 물건을 거래해보세요" />
-    <meta property="og:url" content="https://tokyun02.github.io/" />
-    <title>판다마켓</title>
-    <link rel="icon" href="img/favicon.svg" type="image/svg" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap"
-      rel="stylesheet"
-    />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.min.css"
-    />
-    <link rel="stylesheet" href="style/app.css" />
-  </head>
-
-  <body>
-    <header class="top-navigation">
-      <a href="./index.html"
-        ><img src="img/logo.png" alt="logo img 생성 불가" class="logo"
-      /></a>
-      <a class="login" href="login.html">로그인</a>
-    </header>
-
-    <main>
-      <section class="banner top-banner">
-        <div class="banner-explain">
-          <h1>판다마켓을 소개합니다.</h1>
-          <h2>
-            일상의 모든 물건을<br class="top-banner-title-next-line" />
-            거래해 보세요
-          </h2>
-          <a class="browsing" href="items.html">구경하러 가기</a>
-        </div>
-        <img
-          src="img/Img_home_top.png"
-          alt="top banner img 생성 불가"
-          class="top-banner-img"
-        />
-      </section>
-
-      <section class="home-main-section">
-        <img
-          src="img/img_home_01.png"
-          alt="section1 img 생성 불가"
-          class="section-img"
-        />
-        <div class="section-explain">
-          <span class="keyword">Hot item</span>
-          <h2 class="section-title">
-            인기 상품을<br class="section-title-next-line" />
-            확인해 보세요
-          </h2>
-          <p>
-            가장 HOT한 중고거래 물품을<br />
-            판다 마켓에서 확인해 보세요
-          </p>
-        </div>
-      </section>
-
-      <section class="home-main-section">
-        <div class="section-explain">
-          <span class="keyword">Search</span>
-          <h2 class="section-title">
-            구매를 원하는<br class="section-title-next-line" />
-            상품을 검색하세요
-          </h2>
-          <p>
-            구매하고 싶은 물품은 검색해서<br />
-            쉽게 찾아보세요
-          </p>
-        </div>
-        <img
-          src="img/Img_home_02.png"
-          alt="section2 img 생성 불가"
-          class="section-img"
-        />
-      </section>
-
-      <section class="home-main-section">
-        <img
-          src="img/Img_home_03.png"
-          alt="section3 img 생성 불가"
-          class="section-img"
-        />
-        <div class="section-explain">
-          <span class="keyword">Register</span>
-          <h2 class="section-title">
-            판매를 원하는<br class="section-title-next-line" />
-            상품을 등록하세요
-          </h2>
-          <p>
-            어떤 물건이든 판매하고 싶은 상품을<br />
-            쉽게 등록하세요
-          </p>
-        </div>
-      </section>
-
-      <section class="banner bottom-banner">
-        <div class="banner-explain">
-          <h2>
-            믿을 수 있는<br />
-            판다마켓 중고 거래
-          </h2>
-        </div>
-
-        <img
-          src="img/Img_home_bottom.png"
-          alt="bottom banner img 생성 불가"
-          class="bottom-banner-img"
-        />
-      </section>
-    </main>
-
-    <footer>
-      <span class="creater-info">©Codeit - 2024</span>
-      <ul class="support-list">
-        <li class="support-list-item">
-          <a href="privacy.html" class="support-text privacy">Privacy Policy</a>
-        </li>
-        <li class="support-list-item">
-          <a href="faq.html" class="support-text">FAQ</a>
-        </li>
-      </ul>
-
-      <ul class="social-list">
-        <li class="social-list-item">
-          <a href="https://www.facebook.com" target="_blank" rel="noreferrer"
-            ><img
-              src="img/ic_facebook.png"
-              alt="facebook img 생성 불가"
-              class="social-img"
-          /></a>
-        </li>
-        <li class="social-list-item">
-          <a href="https://www.x.com" target="_blank" rel="noreferrer"
-            ><img
-              src="img/ic_twitter.png"
-              alt="twitter img 생성 불가"
-              class="social-img"
-          /></a>
-        </li>
-        <li class="social-list-item">
-          <a href="https://www.youtube.com" target="_blank" rel="noreferrer"
-            ><img
-              src="img/ic_youtube.png"
-              alt="youtube img 생성 불가"
-              class="social-img"
-          /></a>
-        </li>
-        <li class="social-list-item">
-          <a href="https://www.instagram.com" target="_blank" rel="noreferrer"
-            ><img
-              src="img/ic_instagram.png"
-              alt="instagram img 생성 불가"
-              class="social-img"
-          /></a>
-        </li>
-      </ul>
-    </footer>
-  </body>
-</html>
