--- login.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <title>판다마켓</title>
-    <link rel="icon" href="img/favicon.svg" type="image/svg" />
-
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap"
-      rel="stylesheet"
-    />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.min.css"
-    />
-    <link rel="stylesheet" href="style/account.css" />
-  </head>
-  <body>
-    <header>
-      <a href="./index.html"
-        ><img src="img/logo.png" alt="logo img 생성불가" class="logo"
-      /></a>
-    </header>
-    <main>
-      <form class="login-info auth-form" method="post" action="./items.html">
-        <label for="email" class="email">이메일</label>
-        <div class="input-id">
-          <input
-            type="email"
-            name="email"
-            id="email"
-            placeholder="이메일을 입력해주세요"
-            required
-            aria-required="true"
-          />
-        </div>
-
-        <label for="password" class="password">비밀번호</label>
-        <div class="input-password">
-          <input
-            type="password"
-            name="password"
-            id="password"
-            placeholder="비밀번호를 입력해주세요"
-            required
-            aria-required="true"
-          />
-          <img
-            src="img/img_closed_eye.png"
-            alt="closed_eye img 생성불가"
-            class="toggle-password"
-          />
-        </div>
-
-        <button type="submit" class="login-btn submitButton" disabled>
-          로그인
-        </button>
-      </form>
-
-      <nav class="simple-login">
-        <p class="simple-login-text">간편 로그인하기</p>
-        <div class="login-list">
-          <a href="https://www.google.com/" target="_blank" rel="noreferrer"
-            ><img src="img/ic_google.png" alt="google img 생성불가"
-          /></a>
-          <a
-            href="https://www.kakaocorp.com/page/"
-            target="_blank"
-            rel="noreferrer"
-            ><img src="img/ic_kakao.png" alt="kakao img 생성불가"
-          /></a>
-        </div>
-      </nav>
-    </main>
-    <footer>
-      <span>판다마켓이 처음이신가요? </span>
-      <a href="./signup.html">회원가입</a>
-    </footer>
-    <script src="./scripts/account.js" type="module"></script>
-  </body>
-</html>
