+++ src/App.jsx
@@ -1,0 +1,11 @@
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+
+function App() {
+  return (
+    <>
+      <h1>리액트 프로젝트 시작</h1>
+    </>
+  );
+}
+
+export default App;
