--- style/account.css
@@ -1,192 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-:root {
-  --blue: #3692ff;
-  --gray900: #111827;
-  --gray800: #1f2937;
-  --gray700: #374151;
-  --gray600: #4b5563;
-  --gray500: #6b7280;
-  --gray400: #9ca3af;
-  --gray200: #e5e7eb;
-  --gray100: #f3f4f6;
-  --gray50: #f9fafb;
-}
-
-body {
-  font-family: "Noto Sans KR", sans-serif;
-}
-
-header {
-  margin-top: 40px;
-}
-
-header,
-footer {
-  text-align: center;
-}
-
-main {
-  width: 640px;
-  margin: 0 auto;
-}
-
-label,
-button {
-  display: block;
-}
-
-label {
-  margin: 16px 0;
-  color: var(--gray800);
-  font-size: 18px;
-  font-weight: 700;
-}
-
-input {
-  width: 640px;
-  height: 56px;
-  padding: 24px 24px;
-  border: 2px solid var(--gray100);
-  border-radius: 14px;
-  background-color: var(--gray100);
-  outline: none;
-}
-
-input:focus {
-  border: 2px solid var(--blue);
-  border-radius: 14px;
-}
-
-button {
-  margin-top: 16px;
-}
-
-footer {
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 24px;
-  margin-bottom: 50px;
-}
-
-.login-info {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.logo {
-  width: 396px;
-}
-
-input::placeholder {
-  color: var(--gray400);
-}
-
-.input-password,
-.input-password-confirm {
-  position: relative;
-}
-
-.toggle-password {
-  position: absolute;
-  top: 14px;
-  right: 14px;
-  cursor: pointer;
-}
-
-.login-btn {
-  padding: 16px 0;
-  border: 1px solid var(--blue);
-  border-radius: 40px;
-  background-color: var(--blue);
-  color: var(--gray100);
-  font-size: 20px;
-  font-weight: 600;
-  line-height: 32px;
-  cursor: pointer;
-}
-
-.login-btn:disabled {
-  background-color: #9ca3af;
-  border-color: #9ca3af;
-  pointer-events: none;
-}
-
-.login-btn:hover,
-.login-btn:focus {
-  background-color: #1967d6;
-  border-color: #1967d6;
-}
-
-.simple-login {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 24px 0;
-  padding: 16px 16px;
-  border: 1px solid #e6f2ff;
-  border-radius: 8px;
-  background-color: #e6f2ff;
-}
-
-.error-message {
-  display: inline-block;
-  margin-top: 8px;
-  padding-left: 16px;
-  color: #f74747;
-  font-size: 14px;
-  font-weight: 600;
-}
-
-.error-input-border {
-  border: 1px solid red;
-  border-radius: 14px;
-}
-
-/* tablet_size*/
-@media (max-width: 1199px) {
-  /* 현재까진 타블렛과 pc의 요소 사이즈를 동일하게 설계함 */
-}
-/* mobile_size */
-@media (max-width: 767px) {
-  header,
-  main,
-  footer {
-    max-width: 400px;
-    padding-left: 16px;
-    padding-right: 16px;
-    margin: 0 auto;
-    width: 100vw;
-  }
-  form,
-  label,
-  input,
-  nav,
-  .login-btn,
-  .register-btn {
-    max-width: calc(400px - 32px);
-    width: 90vw;
-  }
-  label {
-    font-size: 14px;
-  }
-  .logo {
-    max-width: calc(400px - 32px);
-    width: 50vw;
-    margin-top: 5vw;
-  }
-  .login-btn,
-  .register-btn {
-    font-size: 16px;
-    padding: 10px 0;
-  }
-  .simple-login-text {
-    flex-shrink: 0;
-  }
-  .login-list {
-    text-align: right;
-  }
-}
