--- style/app.css
@@ -1,323 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-:root {
-  --blue: #3692ff;
-  --gray900: #111827;
-  --gray800: #1f2937;
-  --gray700: #374151;
-  --gray600: #4b5563;
-  --gray500: #6b7280;
-  --gray400: #9ca3af;
-  --gray200: #e5e7eb;
-  --gray100: #f3f4f6;
-  --gray50: #f9fafb;
-}
-
-body {
-  font-family: "Noto Sans KR", sans-serif;
-}
-
-h1 {
-  visibility: hidden;
-}
-
-h2 {
-  color: var(--gray700);
-  font-size: 40px;
-  font-weight: 700;
-  line-height: 56px;
-}
-
-p {
-  color: var(--gray700);
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 32px;
-}
-
-a {
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.home-main-section {
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.top-navigation {
-  display: flex;
-  position: sticky;
-  top: 0;
-  left: 0;
-  justify-content: space-around;
-  align-items: center;
-  margin-bottom: 10px;
-  padding-top: 10px;
-  background-color: #ffffff;
-}
-
-.logo {
-  width: 153px;
-  height: 51px;
-}
-
-.login {
-  padding: 16px 36px;
-  border-radius: 8px;
-  background-color: var(--blue);
-  color: #ffffff;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.banner {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-top: 100px;
-  background-color: #cfe5ff;
-}
-
-.browsing {
-  display: block;
-  width: 357px;
-  height: 56px;
-  margin: 16px 0;
-  padding: 16px 124px;
-  border-radius: 40px;
-  background-color: var(--blue);
-  color: #ffffff;
-  text-align: center;
-}
-
-.top-banner-img {
-  width: 744px;
-  height: 340px;
-}
-
-.home-main-section {
-  display: flex;
-  justify-content: center;
-  gap: 5%;
-  padding: 138px 0;
-  text-align: left;
-}
-
-.home-main-section:nth-of-type(3) {
-  text-align: right;
-}
-
-.section-explain {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  background-color: #fcfcfc;
-}
-
-.home-main-section:nth-of-type(3) .section-explain {
-  padding-right: 40px;
-  padding-left: 0;
-}
-
-.section-img {
-  width: 579px;
-  height: 444px;
-}
-
-.keyword {
-  color: var(--blue);
-  font-size: 18px;
-  font-weight: 700;
-  line-height: 26px;
-}
-
-.bottom-banner {
-  margin-top: 36px;
-}
-
-.bottom-banner-img {
-  width: 744px;
-  height: 397px;
-}
-
-footer {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  padding-top: 30px;
-  padding-bottom: 90px;
-  background-color: var(--gray900);
-}
-
-.creater-info {
-  color: var(--gray400);
-  font-size: 16px;
-  font-weight: 400;
-}
-
-.support-text {
-  color: var(--gray200);
-  font-size: 16px;
-  font-weight: 400;
-}
-
-.social-img {
-  width: 18px;
-  height: 18px;
-}
-
-.support-list,
-.social-list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.support-list-item,
-.social-list-item {
-  display: inline-block;
-}
-
-.privacy {
-  padding-right: 16px;
-}
-/* wide_screen_monitor(1200px~1920px)*/
-@media (max-width: 1920px) {
-  .top-navigation {
-    justify-content: space-between;
-    padding-left: 200px;
-    padding-right: 200px;
-  }
-}
-/* Tablet_screen */
-@media (max-width: 1199px) {
-  h2 {
-    width: 100%;
-    font-size: 40px;
-    font-weight: 700;
-    line-height: 40px;
-    text-align: center;
-  }
-  p {
-    font-size: 14px;
-    font-weight: 500;
-    line-height: 24px;
-  }
-
-  .top-navigation {
-    margin: 0;
-    padding: 16px 24px;
-  }
-
-  .banner {
-    flex-direction: column;
-    padding: 30px 0 0;
-  }
-  .banner-explain {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .top-banner-img,
-  .bottom-banner-img {
-    margin-top: 100px;
-    width: 100%;
-    height: 60%;
-  }
-
-  .top-banner-title-next-line,
-  .section-title-next-line {
-    display: none;
-  }
-
-  .bottom-banner {
-    margin-top: 0;
-    padding-top: 40px;
-  }
-
-  .home-main-section {
-    flex-direction: column;
-    align-items: flex-start;
-    margin: 0;
-    padding: 16px 24px;
-    width: 100%;
-  }
-  .home-main-section:nth-of-type(3) {
-    flex-direction: column-reverse;
-    align-items: flex-end;
-  }
-  .home-main-section h2 {
-    margin: 16px 0;
-    text-align: left;
-  }
-  .home-main-section:nth-of-type(3) h2 {
-    text-align: right;
-  }
-  .home-main-section:nth-of-type(3) .section-explain {
-    padding: 0;
-  }
-  .section-img {
-    width: 100%;
-    height: 70%;
-  }
-  .section-title {
-    font-size: 32px;
-  }
-
-  .browsing {
-    font-size: 16px;
-  }
-
-  .keyword {
-    margin-top: 16px;
-  }
-}
-/* mobile_size */
-@media (max-width: 767px) {
-  h1 {
-    margin: 0;
-    height: 0;
-  }
-  h2 {
-    font-size: 32px;
-  }
-  p {
-    margin: 0 0 40px;
-  }
-  footer {
-    flex-wrap: wrap;
-    justify-content: space-between;
-    align-items: center;
-    padding: 32px;
-    gap: 60px;
-  }
-  .top-navigation {
-    padding: 16px;
-  }
-
-  .top-banner-title-next-line {
-    display: inline;
-  }
-
-  .browsing {
-    padding: 16px 32px;
-    width: 50%;
-  }
-  .home-main-section {
-    padding: 16px;
-  }
-  .section-title {
-    font-size: 22px;
-  }
-  .creater-info {
-    flex-basis: 100%;
-    order: 3;
-  }
-  .social-list-item {
-    padding-right: 10px;
-  }
-}
